backend/requirements.txt
fastapi
uvicorn
pandas
pyyaml
pydantic
duckdb
typer
orjson
numpy

